+++ packages/prerenderer/jest.config.js
@@ -1,0 +1,4 @@
+module.exports = {
+  testMatch: ['**/__tests__/**/*.?(m)[jt]s?(x)', '**/?(*.)+(spec|test).?(m)[tj]s?(x)'],
+  transform: {},
+}