+++ packages/prerenderer/test/mjs.test.mjs
@@ -1,0 +1,6 @@
+import Prerenderer from '@prerenderer/prerenderer'
+
+test('it imports correctly in module', () => {
+  expect(typeof Prerenderer).toEqual('function')
+  expect('default' in Prerenderer).toBeFalsy()
+})