+++ packages/renderer-puppeteer/index.mjs
@@ -1,0 +1,3 @@
+import module from './dist/cjs.js'
+
+export default module