+++ packages/renderer-jsdom/CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.1.0] - 2023-03-23
+
+### Changed
+- Upgrade typescript to v5
+- fix broken mjs support (`.default` was needed when importing)
+- `renderAfterTime`, `renderAfterElementExists` and `renderAfterEvent` can now run in parallel, the first to resolve will render the page
+
+## [1.0.3] - 2023-01-21
+
 ### Changed
 - Monorepo migration
 - Typescript migration